refactor(shared): rename asset_record_object.py to asset_record.py
値オブジェクトのファイル名から *_object.py サフィックスを除去。
クラス名 AssetRecord は変更なし。テストファイル
test_asset_record_object.py も同名規則で追従リネーム。
