+++ lambda/shared/src/shared/domain/asset_record.py
