+++ lambda/web-scraping/tests/domain/test_asset_record.py
